requirements.txt
PyGithub==2.3.0
better_profanity